+++ spring-cloud-starter-tencent-polaris-circuitbreaker/src/test/java/com/tencent/cloud/polaris/circuitbreaker/config/PolarisCircuitBreakerAutoConfigurationTest.java
@@ -20,7 +20,6 @@ package com.tencent.cloud.polaris.circuitbreaker.config;
 import com.tencent.cloud.polaris.circuitbreaker.common.CircuitBreakerConfigModifier;
 import com.tencent.cloud.polaris.context.config.PolarisContextAutoConfiguration;
 import com.tencent.cloud.rpc.enhancement.config.RpcEnhancementAutoConfiguration;
-import com.tencent.polaris.circuitbreak.api.CircuitBreakAPI;
 import org.junit.jupiter.api.Test;
 
 import org.springframework.boot.autoconfigure.AutoConfigurations;
@@ -61,7 +60,6 @@ public class PolarisCircuitBreakerAutoConfigurationTest {
 			assertThat(context).hasSingleBean(PolarisCircuitBreakerAutoConfiguration.class);
 			assertThat(context).hasSingleBean(CircuitBreakerFactory.class);
 			assertThat(context).hasSingleBean(CircuitBreakerConfigModifier.class);
-			assertThat(context).hasSingleBean(CircuitBreakAPI.class);
 			assertThat(context).hasSingleBean(CircuitBreakerNameResolver.class);
 		});
 	}
@@ -72,10 +70,7 @@ public class PolarisCircuitBreakerAutoConfigurationTest {
 			assertThat(context).hasSingleBean(ReactivePolarisCircuitBreakerAutoConfiguration.class);
 			assertThat(context).hasSingleBean(ReactiveCircuitBreakerFactory.class);
 			assertThat(context).hasSingleBean(CircuitBreakerConfigModifier.class);
-			assertThat(context).hasSingleBean(CircuitBreakAPI.class);
 		});
 	}
-
-
 
 }