+++ spring-cloud-starter-tencent-polaris-discovery/src/main/java/com/tencent/cloud/polaris/discovery/refresh/PolarisRefreshApplicationReadyEventListener.java
@@ -21,7 +21,7 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
 
-import com.tencent.cloud.polaris.discovery.PolarisDiscoveryHandler;
+import com.tencent.cloud.polaris.context.PolarisSDKContextManager;
 import com.tencent.polaris.client.util.NamedThreadFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -45,14 +45,14 @@ public class PolarisRefreshApplicationReadyEventListener
 
 	private static final Logger LOG = LoggerFactory.getLogger(PolarisRefreshApplicationReadyEventListener.class);
 	private static final int DELAY = 60;
-	private final PolarisDiscoveryHandler polarisDiscoveryHandler;
+	private final PolarisSDKContextManager polarisSDKContextManager;
 	private final PolarisServiceStatusChangeListener polarisServiceStatusChangeListener;
 	private final ScheduledExecutorService refreshExecutor;
 	private ApplicationEventPublisher publisher;
 
-	public PolarisRefreshApplicationReadyEventListener(PolarisDiscoveryHandler polarisDiscoveryHandler,
+	public PolarisRefreshApplicationReadyEventListener(PolarisSDKContextManager polarisSDKContextManager,
 			PolarisServiceStatusChangeListener polarisServiceStatusChangeListener) {
-		this.polarisDiscoveryHandler = polarisDiscoveryHandler;
+		this.polarisSDKContextManager = polarisSDKContextManager;
 		this.polarisServiceStatusChangeListener = polarisServiceStatusChangeListener;
 		this.refreshExecutor = Executors.newSingleThreadScheduledExecutor(
 				new NamedThreadFactory("polaris-service-refresh"));
@@ -61,7 +61,7 @@ public class PolarisRefreshApplicationReadyEventListener
 	@Override
 	public void onApplicationEvent(ApplicationReadyEvent event) {
 		// Register service change listener.
-		polarisDiscoveryHandler.getSdkContext().getExtensions().getLocalRegistry()
+		polarisSDKContextManager.getSDKContext().getExtensions().getLocalRegistry()
 				.registerResourceListener(polarisServiceStatusChangeListener);
 
 		// Begin scheduled refresh thread.