+++ spring-cloud-starter-tencent-polaris-discovery/src/main/java/com/tencent/cloud/polaris/registry/PolarisServiceRegistry.java
@@ -26,6 +26,7 @@ import java.util.concurrent.ScheduledExecutorService;
 
 import com.tencent.cloud.common.metadata.StaticMetadataManager;
 import com.tencent.cloud.polaris.PolarisDiscoveryProperties;
+import com.tencent.cloud.polaris.context.PolarisSDKContextManager;
 import com.tencent.cloud.polaris.discovery.PolarisDiscoveryHandler;
 import com.tencent.cloud.polaris.util.OkHttpUtil;
 import com.tencent.cloud.rpc.enhancement.stat.config.PolarisStatProperties;
@@ -47,6 +48,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import org.springframework.beans.BeanUtils;
+import org.springframework.beans.factory.DisposableBean;
 import org.springframework.cloud.client.serviceregistry.ServiceRegistry;
 import org.springframework.http.HttpHeaders;
 
@@ -58,24 +60,25 @@ import static org.springframework.util.ReflectionUtils.rethrowRuntimeException;
  *
  * @author Haotian Zhang, Andrew Shan, Jie Cheng, changjin wei(魏昌进)
  */
-public class PolarisServiceRegistry implements ServiceRegistry<PolarisRegistration> {
+public class PolarisServiceRegistry implements ServiceRegistry<PolarisRegistration>, DisposableBean {
 
 	private static final Logger LOGGER = LoggerFactory.getLogger(PolarisServiceRegistry.class);
 
 	private final PolarisDiscoveryProperties polarisDiscoveryProperties;
 
+	private final PolarisSDKContextManager polarisSDKContextManager;
+
 	private final PolarisDiscoveryHandler polarisDiscoveryHandler;
 
 	private final StaticMetadataManager staticMetadataManager;
-
+	private final PolarisStatProperties polarisStatProperties;
 	private final ScheduledExecutorService heartbeatExecutor;
 
-	private final PolarisStatProperties polarisStatProperties;
-
 	public PolarisServiceRegistry(PolarisDiscoveryProperties polarisDiscoveryProperties,
-			PolarisDiscoveryHandler polarisDiscoveryHandler,
+			PolarisSDKContextManager polarisSDKContextManager, PolarisDiscoveryHandler polarisDiscoveryHandler,
 			StaticMetadataManager staticMetadataManager, PolarisStatProperties polarisStatProperties) {
 		this.polarisDiscoveryProperties = polarisDiscoveryProperties;
+		this.polarisSDKContextManager = polarisSDKContextManager;
 		this.polarisDiscoveryHandler = polarisDiscoveryHandler;
 		this.staticMetadataManager = staticMetadataManager;
 
@@ -116,7 +119,7 @@ public class PolarisServiceRegistry implements ServiceRegistry<PolarisRegistrati
 		instanceRegisterRequest.setVersion(polarisDiscoveryProperties.getVersion());
 		instanceRegisterRequest.setInstanceId(polarisDiscoveryProperties.getInstanceId());
 		try {
-			ProviderAPI providerClient = polarisDiscoveryHandler.getProviderAPI();
+			ProviderAPI providerClient = polarisSDKContextManager.getProviderAPI();
 			InstanceRegisterResponse instanceRegisterResponse;
 			if (StringUtils.isBlank(polarisDiscoveryProperties.getHealthCheckUrl())) {
 				instanceRegisterResponse = providerClient.registerInstance(instanceRegisterRequest);
@@ -134,7 +137,7 @@ public class PolarisServiceRegistry implements ServiceRegistry<PolarisRegistrati
 					staticMetadataManager.getMergedStaticMetadata());
 			if (Objects.nonNull(polarisStatProperties) && polarisStatProperties.isEnabled()) {
 				try {
-					StatReporter statReporter = (StatReporter) polarisDiscoveryHandler.getSdkContext().getPlugins()
+					StatReporter statReporter = (StatReporter) polarisSDKContextManager.getSDKContext().getPlugins()
 							.getPlugin(PluginTypes.STAT_REPORTER.getBaseType(), StatReporterConfig.DEFAULT_REPORTER_PROMETHEUS);
 					if (Objects.nonNull(statReporter)) {
 						LOGGER.info("Stat server started on port: " + statReporter.metaInfo().getPort() + " (http)");
@@ -148,10 +151,12 @@ public class PolarisServiceRegistry implements ServiceRegistry<PolarisRegistrati
 				}
 			}
 
-			ServiceConfigImpl serviceConfig = (ServiceConfigImpl) polarisDiscoveryHandler.getSdkContext().getConfig()
+			ServiceConfigImpl serviceConfig = (ServiceConfigImpl) polarisSDKContextManager.getSDKContext().getConfig()
 					.getProvider().getService();
 			serviceConfig.setNamespace(polarisDiscoveryProperties.getNamespace());
 			serviceConfig.setName(serviceId);
+
+			PolarisSDKContextManager.isRegistered = true;
 		}
 		catch (Exception e) {
 			LOGGER.error("polaris registry, {} register failed...{},", registration.getServiceId(), registration, e);
@@ -176,7 +181,7 @@ public class PolarisServiceRegistry implements ServiceRegistry<PolarisRegistrati
 		deRegisterRequest.setPort(registration.getPort());
 
 		try {
-			ProviderAPI providerClient = polarisDiscoveryHandler.getProviderAPI();
+			ProviderAPI providerClient = polarisSDKContextManager.getProviderAPI();
 			providerClient.deRegister(deRegisterRequest);
 		}
 		catch (Exception e) {
@@ -186,8 +191,9 @@ public class PolarisServiceRegistry implements ServiceRegistry<PolarisRegistrati
 			if (null != heartbeatExecutor) {
 				heartbeatExecutor.shutdown();
 			}
+			LOGGER.info("De-registration finished.");
+			PolarisSDKContextManager.isRegistered = false;
 		}
-		LOGGER.info("De-registration finished.");
 	}
 
 	@Override
@@ -205,7 +211,7 @@ public class PolarisServiceRegistry implements ServiceRegistry<PolarisRegistrati
 		String serviceName = registration.getServiceId();
 		InstancesResponse instancesResponse = polarisDiscoveryHandler.getInstances(serviceName);
 		Instance[] instances = instancesResponse.getInstances();
-		if (null == instances || instances.length == 0) {
+		if (null == instances) {
 			return null;
 		}
 		for (Instance instance : instances) {
@@ -242,7 +248,7 @@ public class PolarisServiceRegistry implements ServiceRegistry<PolarisRegistrati
 					return;
 				}
 
-				polarisDiscoveryHandler.getProviderAPI().heartbeat(heartbeatRequest);
+				polarisSDKContextManager.getProviderAPI().heartbeat(heartbeatRequest);
 				LOGGER.trace("Polaris heartbeat is sent");
 			}
 			catch (PolarisException e) {
@@ -252,5 +258,12 @@ public class PolarisServiceRegistry implements ServiceRegistry<PolarisRegistrati
 				LOGGER.error("polaris heartbeat runtime error", e);
 			}
 		}, polarisDiscoveryProperties.getHeartbeatInterval(), polarisDiscoveryProperties.getHeartbeatInterval(), SECONDS);
+	}
+
+	@Override
+	public void destroy() {
+		if (heartbeatExecutor != null) {
+			heartbeatExecutor.shutdown();
+		}
 	}
 }